Word separation tuning. Massive char extraction
Word separation is tested on one of the images (90% good). It then saves
words and is able to split them into proper charcters. Now opencv.py
needs to be ported to Python3. This is needed because other files are
running on Python3 and support for Python2 is running out in 2 years.